fix(#2394): retire mockups backend surface — router, models, schemas (BE/EE half of #2378)
FE was fully retired in #2378/#2770 (zero entry points remain, PO's dev
measurement found the mockups table doesn't exist and scenarios/versions/
pages are empty). This removes the corresponding backend surface:

- app/routers/mockups.py (12 routes: CRUD + versions + scenarios) — deleted
- app/models/mockup.py (MockupPage/Component/Scenario/Version) — deleted
- app/schemas/mockup.py — deleted
- app/main.py — unregistered

Extracted, not deleted: GET /api/v2/usage + the UsageMeter model/schema.
They lived in mockups.py/mockup.py but aren't mockups-domain — the
meter_type CHECK constraint doesn't even include 'mockups' as a value.
Deleting them as a side effect of an unrelated cleanup would be scope
creep in the wrong direction. Moved to their own
app/routers/usage.py + app/models/usage_meter.py + app/schemas/usage.py.

test_authz_project_scope_coverage.py — required edit, not optional: its
_ID_MUTATION_FALSE_POSITIVE_ALLOWLIST hardcoded 4 mockups routes, and
test_id_mutation_baseline_entries_are_not_stale asserts every baseline
entry resolves to a live route. Verified via positive control (temporarily
reintroduced one stale entry, confirmed the assertion catches it with the
exact "rot된 라우트" failure it's designed to produce).

Also removed the now-dead mockup_pages entry from purge_test_agents.py's
DELETE_SPECS (had a _table_exists guard so it wasn't actively broken, just
stale — hygiene).

Split test_s48.py: its one GET /usage test moved to a new test_usage.py
(unchanged usage semantics); the rest (mockup CRUD) deleted along with the
router it tested. test_2181_mockup_model_db_drift.py (mockup-specific
column-drift regression test) deleted outright — no longer applicable.

Left alone, out of scope:
- entity_slug.py's reserved-slug denylist still includes "mockups" —
  harmless to leave (workspace slugs just stay slightly over-reserved),
  not worth touching in this PR.
- No DB migration to drop the mockup_* tables. PO's dev measurement found
  them already absent in dev, but nothing in this repo (alembic or the
  Supabase migration chain) ever dropped them — unexplained drift, not a
  tracked removal. Whether they still exist in prod is unverified (no prod
  access this session) — if they do, an actual DROP TABLE migration would
  need to be authored separately with infra/PO sign-off, not assumed away.

max_mockups EE billing seed (plan_features rows, Free tier limit_value=5,
Team/Pro unlimited) — reported as facts only in the PR body per PO's
explicit instruction, not touched here. It's pure seed/config data with no
backend enforcement code reading it (confirmed via grep — no Python
references plan_features.max_mockups at all).
