--- backend/app/models/mockup.py
@@ -1,79 +1,0 @@
-import uuid
-from datetime import datetime
-
-from sqlalchemy import Boolean, DateTime, Integer, Text, func
-from sqlalchemy.dialects.postgresql import JSONB, UUID
-from sqlalchemy.orm import Mapped, mapped_column
-
-from app.core.database import Base
-
-
-class MockupPage(Base):
-    __tablename__ = "mockup_pages"
-
-    id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    org_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
-    project_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
-    slug: Mapped[str] = mapped_column(Text, nullable=False)
-    title: Mapped[str] = mapped_column(Text, nullable=False)
-    category: Mapped[str | None] = mapped_column(Text, nullable=True)
-    # story #2181(모델↔DB drift 감사): 실 DB는 NOT NULL(모델은 nullable=True로 잘못 선언돼 있었음
-    # — router가 body.viewport 그대로 넘기면 생략 호출 시 NotNullViolation 500이던 자리, 같이 fix).
-    viewport: Mapped[str] = mapped_column(Text, nullable=False, server_default="desktop")
-    version: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    created_by: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
-    updated_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
-    deleted_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
-
-
-class MockupComponent(Base):
-    __tablename__ = "mockup_components"
-
-    id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    page_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
-    # story #2181(모델↔DB drift 감사, 2026-07-24): 실 DB 컬럼명은 `component_type`(모델은
-    # `type`이라 선언 — 처음부터 한 번도 일치한 적 없었던 자리, mockup_components를 건드리는
-    # 모든 select(MockupComponent)가 실 migrated DB에서 UndefinedColumnError였을 것). API
-    # 계약(MockupComponentOut.type)·라우터 kwarg(`type=`)는 그대로 두고 DB 컬럼명만 명시
-    # 매핑(app/models/agent_run.py의 run_metadata↔"metadata"와 동형 패턴) — 파이썬 쪽 이름은
-    # 안 건드려 회귀 0.
-    type: Mapped[str] = mapped_column("component_type", Text, nullable=False)
-    props: Mapped[dict] = mapped_column(JSONB, nullable=False, default=dict)
-    sort_order: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    parent_id: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
-
-
-class MockupScenario(Base):
-    __tablename__ = "mockup_scenarios"
-
-    id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    page_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
-    name: Mapped[str] = mapped_column(Text, nullable=False)
-    override_props: Mapped[dict] = mapped_column(JSONB, nullable=False, default=dict)
-    is_default: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    sort_order: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
-
-
-class MockupVersion(Base):
-    __tablename__ = "mockup_versions"
-
-    id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    page_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
-    version: Mapped[int] = mapped_column(Integer, nullable=False)
-    snapshot: Mapped[dict] = mapped_column(JSONB, nullable=False, default=dict)
-    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
-
-
-class UsageMeter(Base):
-    __tablename__ = "usage_meters"
-
-    id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    org_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
-    meter_type: Mapped[str] = mapped_column(Text, nullable=False)
-    current_value: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    limit_value: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    period_start: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
-    period_end: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
