+++ backend/app/models/usage_meter.py
@@ -1,0 +1,20 @@
+import uuid
+from datetime import datetime
+
+from sqlalchemy import DateTime, Integer, Text
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.orm import Mapped, mapped_column
+
+from app.core.database import Base
+
+
+class UsageMeter(Base):
+    __tablename__ = "usage_meters"
+
+    id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
+    org_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=False, index=True)
+    meter_type: Mapped[str] = mapped_column(Text, nullable=False)
+    current_value: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    limit_value: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    period_start: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
+    period_end: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
